tail follow-up
I was too tired to color tail when I published it yesterday, but I
regret doing so. the new GIMP gif file is too big and the old one is
outdated. if you'd like a full 2400x1600 (300dpi) gimp file of the gif
then email me at zerglingsascats@gmail.com and I'll find a way to get
it to you :)